Fix the Hotwire stuff by renaming in templates
