+++ app/views/cbv_flows/employer_search.html.erb
@@ -5,12 +5,12 @@
 
 <h2>Get payment info from your employer.</h2>
 
-<form data-controller="providers" action="<%= next_path %>" method="get">
+<form data-controller="cbv-flows" action="<%= next_path %>" method="get">
   <label class="usa-label" for="employer">Search for an employer</label>
-  <div class="usa-combo-box margin-bottom-3" data-action="input->providers#search">
-    <select class="usa-select" name="employer" id="employer" data-providers-target="options" data-action="change->providers#select">
+  <div class="usa-combo-box margin-bottom-3" data-action="input->cbv-flows#search">
+    <select class="usa-select" name="employer" id="employer" data-cbv-flows-target="options" data-action="change->cbv-flows#select">
       <option value>Select an employer</option>
     </select>
   </div>
-  <button data-action="click->providers#submit" class="usa-button usa-button--outline" disabled="disabled" type="submit" data-providers-target="continue">Continue</button>
+  <button data-action="click->cbv-flows#submit" class="usa-button usa-button--outline" disabled="disabled" type="submit" data-cbv-flows-target="continue">Continue</button>
 </form>
